requirements.txt
Flask==3.0.0
pytest==7.4.0
openai==1.54.3
python-dotenv==1.0.1
mock
pytest-mock
unittest